Clean Gitorious leftovers in Repository
 - Remove most columns from repository that seems like uninteresting
   leftovers from gitorious times.
 - Clean up in the model, both adjusting/removing functions that used
   now removed columns and stuff that is clearly not of use in water.
 - Remove corresponding stuff in the repository test which was
   originally written by the gitorious team.

It was difficult to decide what to remove and what not to, my attitude
was to remove if I felt uncertain. The decision to remove the
when-to-gc tracking stuff was that in Water it's most sane to gc when
the corresponding lhg has been Accepted, since it's then locked for
further changes.